Extract integration tests tied to Ammonite REPL to a separate trait
