+++ modules/integration/src/test/scala/scala/cli/integration/ReplTests3Lts.scala
@@ -1,3 +1,6 @@
 package scala.cli.integration
 
-class ReplTests3Lts extends ReplTestDefinitions with ReplTests3StableDefinitions with Test3Lts
+class ReplTests3Lts extends ReplTestDefinitions
+    with ReplAmmoniteTestDefinitions
+    with ReplAmmoniteTests3StableDefinitions
+    with Test3Lts
